Change color parameter type from BYTE to DWORD

+++ Font/text.h
@@ -1,10 +1,10 @@
-void DrawChar(BYTE* bitmap, BYTE color);
-void Print(const char* str, BYTE color);
+void DrawChar(BYTE* bitmap, DWORD color);
+void Print(const char* str, DWORD color);
 void Debug(const CHAR* str, int debug);
 void IntToString(int value, char* buffer);
-void PrintInt(int value, BYTE color);
-void PrintHex(int value, BYTE color);
-void PrintOut(char letter, BYTE color);
+void PrintInt(int value, DWORD color);
+void PrintHex(int value, DWORD color);
+void PrintOut(char letter, DWORD color);
 
 int atoi(const char* str);
 int GetCursorX();
